Move workflows stuff into submodule
