+++ ironflow/nodes/pyiron/atomistics_nodes.py
@@ -19,7 +19,7 @@ from pyiron_atomistics.atomistics.structure.factory import StructureFactory
 from pyiron_atomistics.atomistics.job.atomistic import AtomisticGenericJob
 from pyiron_atomistics.lammps import list_potentials
 
-from ironflow.gui.canvas_widgets.nodes import ButtonNodeWidget
+from ironflow.gui.workflows.canvas_widgets.nodes import ButtonNodeWidget
 from ironflow.model import dtypes, NodeInputBP, NodeOutputBP
 from ironflow.model.node import Node
 from ironflow.nodes.std.special_nodes import DualNodeBase
